Spells refactor
