+++ src/assets/db/spells.js
@@ -91,7 +91,7 @@ export const spells = [
     ritual: true,
     duration: "8 hrs",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 1,
     school: {
       name: "Abjuration",
@@ -289,7 +289,7 @@ export const spells = [
     ritual: false,
     duration: "Instant",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 3,
     school: {
       name: "Necromancy",
@@ -773,7 +773,7 @@ export const spells = [
     ritual: true,
     duration: "Instant",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 2,
     school: {
       name: "Divination",
@@ -802,7 +802,7 @@ export const spells = [
     ],
     higher_level: [""],
     page: "216",
-    range: "Self (30-foot radius)",
+    range: "Self (30-feet)",
     components: ["V"],
     material: "",
     ritual: false,
@@ -832,7 +832,7 @@ export const spells = [
     ],
     higher_level: [""],
     page: "216",
-    range: "Self (30-foot radius)",
+    range: "Self (30-feet)",
     components: ["V"],
     material: "",
     ritual: false,
@@ -862,7 +862,7 @@ export const spells = [
     ],
     higher_level: [""],
     page: "216",
-    range: "Self (30-foot radius)",
+    range: "Self (30-feet)",
     components: ["V"],
     material: "",
     ritual: false,
@@ -1442,7 +1442,7 @@ export const spells = [
     range: "30 feet",
     components: ["V"],
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 2,
@@ -1491,7 +1491,7 @@ export const spells = [
     range: "Self",
     components: ["V", "S"],
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 3,
@@ -1850,7 +1850,7 @@ export const spells = [
     components: ["V", "S", "M"],
     material: "a diamond worth at least 50 gp",
     ritual: false,
-    duration: "",
+    duration: "Instant",
     concentration: false,
     casting_time: "1 action",
     level: 1,
@@ -1920,7 +1920,7 @@ export const spells = [
     ],
     higher_level: [""],
     page: "221",
-    range: "Self (30-foot radius)",
+    range: "Self (30-feet)",
     components: ["V"],
     material: "",
     ritual: false,
@@ -2212,9 +2212,9 @@ export const spells = [
     components: ["V", "S", "M"],
     material: "Incense and a vial of holy or unholy water.",
     ritual: true,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 5,
     school: {
       name: "Divination",
@@ -2253,7 +2253,7 @@ export const spells = [
     ritual: true,
     duration: "Instant",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 5,
     school: {
       name: "Divination",
@@ -2560,7 +2560,7 @@ export const spells = [
     ritual: false,
     duration: "Up to 1 hr",
     concentration: true,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 7,
     school: {
       name: "Conjuration",
@@ -2595,7 +2595,7 @@ export const spells = [
     ritual: false,
     duration: "Up to 1 hr",
     concentration: true,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 5,
     school: {
       name: "Conjuration",
@@ -2637,7 +2637,7 @@ export const spells = [
     ritual: false,
     duration: "Up to 1 hr",
     concentration: true,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 6,
     school: {
       name: "Conjuration",
@@ -2678,7 +2678,7 @@ export const spells = [
     ritual: false,
     duration: "Up to 1 hr",
     concentration: true,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 4,
     school: {
       name: "Conjuration",
@@ -2780,9 +2780,9 @@ export const spells = [
     range: "Self",
     components: ["V"],
     ritual: true,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 5,
     school: {
       name: "Divination",
@@ -3136,7 +3136,7 @@ export const spells = [
     ritual: false,
     duration: "Instant",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 6,
     school: {
       name: "Necromancy",
@@ -3233,7 +3233,7 @@ export const spells = [
     ritual: false,
     duration: "Special",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 5,
     school: {
       name: "Illusion",
@@ -3683,7 +3683,7 @@ export const spells = [
     ],
     higher_level: [""],
     page: "231",
-    range: "Self (30-foot radius)",
+    range: "Self (30-feet)",
     components: ["V"],
     material: "",
     ritual: false,
@@ -4381,7 +4381,7 @@ export const spells = [
     ritual: true,
     duration: "until dispelled",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 6,
     school: {
       name: "Conjuration",
@@ -4414,7 +4414,7 @@ export const spells = [
     ritual: false,
     duration: "8 hrs",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 5,
     school: {
       name: "Illusion",
@@ -4761,7 +4761,7 @@ export const spells = [
     range: "60 feet",
     components: ["V", "S"],
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 2,
@@ -5112,7 +5112,7 @@ export const spells = [
     components: ["V", "M"],
     material: "A small feather or a piece of down.",
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 reaction",
     level: 1,
@@ -5355,7 +5355,7 @@ export const spells = [
     ritual: false,
     duration: "Up to 24 hrs",
     concentration: true,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 6,
     school: {
       name: "Divination",
@@ -5892,7 +5892,7 @@ export const spells = [
     ritual: false,
     duration: "8 hrs",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 9,
     school: {
       name: "Divination",
@@ -6147,7 +6147,7 @@ export const spells = [
     ritual: false,
     duration: "30 days",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 5,
     school: {
       name: "Enchantment",
@@ -6446,7 +6446,7 @@ export const spells = [
     components: ["V", "S", "M"],
     material: "A bit of pork rind or butter.",
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 1,
@@ -7495,7 +7495,7 @@ export const spells = [
     ritual: true,
     duration: "Instant",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 1,
     school: {
       name: "Divination",
@@ -7536,7 +7536,7 @@ export const spells = [
     ritual: true,
     duration: "10 days",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 1,
     school: {
       name: "Illusion",
@@ -7592,7 +7592,7 @@ export const spells = [
     ritual: false,
     duration: "Until dispelled",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 9,
     school: {
       name: "Abjuration",
@@ -7820,7 +7820,7 @@ export const spells = [
     components: ["V", "S", "M"],
     material: "A grasshopper's hind leg.",
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 1,
@@ -7981,7 +7981,7 @@ export const spells = [
     ritual: true,
     duration: "8 hrs",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 3,
     school: {
       name: "Evocation",
@@ -8473,7 +8473,7 @@ export const spells = [
     range: "30 feet",
     components: ["V", "S"],
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 0,
@@ -8528,7 +8528,7 @@ export const spells = [
     ritual: false,
     duration: "1 hour",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 3,
     school: {
       name: "Abjuration",
@@ -8580,7 +8580,7 @@ export const spells = [
     ritual: false,
     duration: "Until dispelled",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 6,
     school: {
       name: "Necromancy",
@@ -8651,7 +8651,7 @@ export const spells = [
     ritual: true,
     duration: "Until dispelled",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 2,
     school: {
       name: "Illusion",
@@ -9029,7 +9029,7 @@ export const spells = [
     ritual: false,
     duration: "Instant",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 0,
     school: {
       name: "Transmutation",
@@ -9184,7 +9184,7 @@ export const spells = [
     components: ["S", "M"],
     material: "A bit of fleece.",
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 0,
@@ -9270,7 +9270,7 @@ export const spells = [
     range: "Self",
     components: ["V", "S"],
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 2,
@@ -9505,7 +9505,7 @@ export const spells = [
     ritual: false,
     duration: "24 hrs",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 7,
     school: {
       name: "Conjuration",
@@ -9977,7 +9977,7 @@ export const spells = [
     ritual: true,
     duration: "1 hour",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 3,
     school: {
       name: "Illusion",
@@ -11023,7 +11023,7 @@ export const spells = [
     ritual: false,
     duration: "1 hour",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 7,
     school: {
       name: "Transmutation",
@@ -11382,7 +11382,7 @@ export const spells = [
     components: ["V", "S", "M"],
     material: "A small silver mirror.",
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 bonus action",
     level: 1,
@@ -11868,7 +11868,7 @@ export const spells = [
     components: ["V", "S", "M"],
     material: "Mistletoe, a shamrock leaf, and a club or quarterstaff.",
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 bonus action",
     level: 0,
@@ -12063,7 +12063,7 @@ export const spells = [
     components: ["V", "S", "M"],
     material: "A pinch of fine sand, rose petals, or a cricket.",
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 1,
@@ -12432,7 +12432,7 @@ export const spells = [
       "When you cast this spell using a spell slot of 4th level or higher, the damage increases by 1d8 for each slot level above 3rd. "
     ],
     page: "278",
-    range: "Self (15-foot-radius)",
+    range: "Self (15-feet)",
     components: ["V", "S", "M"],
     material: "a holy symbol",
     ritual: false,
@@ -12505,7 +12505,7 @@ export const spells = [
     range: "60 feet",
     components: ["V", "S"],
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 bonus action",
     level: 2,
@@ -12923,7 +12923,7 @@ export const spells = [
     ritual: false,
     duration: "Until dispelled",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 7,
     school: {
       name: "Abjuration",
@@ -13105,7 +13105,7 @@ export const spells = [
     ritual: false,
     duration: "1 round",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 5,
     school: {
       name: "Conjuration",
@@ -13180,7 +13180,7 @@ export const spells = [
     range: "30 feet",
     components: ["V"],
     ritual: false,
-    duration: "1 minute",
+    duration: "1 min",
     concentration: false,
     casting_time: "1 action",
     level: 0,
@@ -13643,7 +13643,7 @@ export const spells = [
     ritual: false,
     duration: "Up to 6 rounds",
     concentration: true,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 8,
     school: {
       name: "Conjuration",
@@ -14177,7 +14177,7 @@ export const spells = [
     ritual: false,
     duration: "8 hrs",
     concentration: false,
-    casting_time: "1 minute",
+    casting_time: "1 min",
     level: 6,
     school: {
       name: "Transmutation",
